Added Cart interface
